Cmmn engine made compatible with CRDB
