+++ modules/flowable-cmmn-engine/src/main/java/org/flowable/cmmn/engine/impl/db/CmmnDbSchemaManager.java
@@ -15,17 +15,22 @@ package org.flowable.cmmn.engine.impl.db;
 import org.flowable.cmmn.engine.CmmnEngineConfiguration;
 import org.flowable.cmmn.engine.impl.util.CommandContextUtil;
 import org.flowable.common.engine.api.FlowableException;
+import org.flowable.common.engine.impl.AbstractEngineConfiguration;
 import org.flowable.common.engine.impl.db.EngineDatabaseConfiguration;
 import org.flowable.common.engine.impl.db.LiquibaseBasedSchemaManager;
 import org.flowable.common.engine.impl.db.LiquibaseDatabaseConfiguration;
 import org.flowable.common.engine.impl.db.SchemaManager;
+import org.flowable.common.engine.impl.interceptor.EngineConfigurationConstants;
+import org.flowable.idm.engine.IdmEngineConfiguration;
 
 public class CmmnDbSchemaManager extends LiquibaseBasedSchemaManager {
 
     public static final String LIQUIBASE_CHANGELOG = "org/flowable/cmmn/db/liquibase/flowable-cmmn-db-changelog.xml";
 
-    public CmmnDbSchemaManager() {
-        super("cmmn", LIQUIBASE_CHANGELOG, CmmnEngineConfiguration.LIQUIBASE_CHANGELOG_PREFIX);
+    public static final String LIQUIBASE_CHANGELOG_CRDB = "org/flowable/cmmn/db/liquibase/flowable-cmmn-db-changelog-crdb.xml";
+
+    public CmmnDbSchemaManager(String changelogFile) {
+        super("cmmn", changelogFile, CmmnEngineConfiguration.LIQUIBASE_CHANGELOG_PREFIX);
     }
 
     @Override
@@ -36,7 +41,24 @@ public class CmmnDbSchemaManager extends LiquibaseBasedSchemaManager {
     public void initSchema() {
         initSchema(CommandContextUtil.getCmmnEngineConfiguration().getDatabaseSchemaUpdate());
     }
-    
+
+    @Override
+    public void initSchema(String databaseSchemaUpdate) {
+        super.initSchema(databaseSchemaUpdate);
+
+        // When the databaseSchemaUpdate is drop-create:
+        // the IDM engine will have done a drop-create due to the configurator running first
+        // The CmmnDbSchemaManager will go next, but it will do a dropAll, dropping the idm tables too.
+        if (AbstractEngineConfiguration.DB_SCHEMA_UPDATE_DROP_CREATE.equals(databaseSchemaUpdate)) {
+            AbstractEngineConfiguration abstractEngineConfiguration = CommandContextUtil.getCmmnEngineConfiguration().getEngineConfigurations()
+                .get(EngineConfigurationConstants.KEY_IDM_ENGINE_CONFIG);
+            if (abstractEngineConfiguration != null) {
+                IdmEngineConfiguration idmEngineConfiguration = (IdmEngineConfiguration) abstractEngineConfiguration;
+                idmEngineConfiguration.getSchemaManager().schemaCreate();
+            }
+        }
+    }
+
     @Override
     public void schemaCreate() {
         try {
