+++ modules/flowable-cmmn-engine/src/main/java/org/flowable/cmmn/engine/test/AbstractFlowableCmmnTestCase.java
@@ -1,0 +1,143 @@
+/* Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.flowable.cmmn.engine.test;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertTrue;
+
+import java.time.Instant;
+import java.time.temporal.ChronoField;
+import java.util.Date;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.flowable.cmmn.api.CmmnHistoryService;
+import org.flowable.cmmn.api.CmmnManagementService;
+import org.flowable.cmmn.api.CmmnRepositoryService;
+import org.flowable.cmmn.api.CmmnRuntimeService;
+import org.flowable.cmmn.api.CmmnTaskService;
+import org.flowable.cmmn.api.runtime.CaseInstance;
+import org.flowable.cmmn.api.runtime.PlanItemInstance;
+import org.flowable.cmmn.engine.CmmnEngine;
+import org.flowable.cmmn.engine.CmmnEngineConfiguration;
+import org.flowable.cmmn.engine.test.impl.CmmnJobTestHelper;
+import org.flowable.cmmn.engine.test.impl.CmmnTestRunner;
+import org.junit.After;
+import org.junit.runner.RunWith;
+
+/**
+ * @author Joram Barrez
+ */
+@RunWith(CmmnTestRunner.class)
+public abstract class AbstractFlowableCmmnTestCase {
+
+    public static CmmnEngine cmmnEngine;
+
+    protected CmmnEngineConfiguration cmmnEngineConfiguration;
+    protected CmmnManagementService cmmnManagementService;
+    protected CmmnRepositoryService cmmnRepositoryService;
+    protected CmmnRuntimeService cmmnRuntimeService;
+    protected CmmnTaskService cmmnTaskService;
+    protected CmmnHistoryService cmmnHistoryService;
+
+    protected String deploymentId;
+
+    @After
+    public void cleanup() {
+        if (deploymentId != null) {
+           cmmnRepositoryService.deleteDeployment(deploymentId, true);
+        }
+    }
+
+    protected void deployOneHumanTaskCaseModel() {
+        deploymentId = cmmnRepositoryService.createDeployment()
+                .addClasspathResource("org/flowable/cmmn/test/one-human-task-model.cmmn")
+                .deploy()
+                .getId();
+    }
+
+    protected CaseInstance deployAndStartOneHumanTaskCaseModel() {
+        deployOneHumanTaskCaseModel();
+        return cmmnRuntimeService.createCaseInstanceBuilder().caseDefinitionKey("oneTaskCase").start();
+    }
+
+    protected void deployOneTaskCaseModel() {
+        deploymentId = cmmnRepositoryService.createDeployment()
+                .addClasspathResource("org/flowable/cmmn/test/one-task-model.cmmn")
+                .deploy()
+                .getId();
+    }
+    
+    protected Date setClockFixedToCurrentTime() {
+        // SQL Server rounds the milliseconds, on order to be stable we set them to 0
+        Date date = Date.from(Instant.now().with(ChronoField.MILLI_OF_SECOND, 0));
+        cmmnEngineConfiguration.getClock().setCurrentTime(date);
+        return date;
+    }
+
+    protected void setClockTo(long epochTime) {
+        setClockTo(new Date(epochTime));
+    }
+    
+    protected void setClockTo(Date date) {
+        cmmnEngineConfiguration.getClock().setCurrentTime(date);
+    }
+
+    protected Date forwardClock(long milliseconds) {
+        long currentMillis = cmmnEngineConfiguration.getClock().getCurrentTime().getTime();
+        Date date = new Date(currentMillis + milliseconds);
+        cmmnEngineConfiguration.getClock().setCurrentTime(date);
+        return date;
+    }
+
+    protected void assertCaseInstanceEnded(CaseInstance caseInstance) {
+        long count = cmmnRuntimeService.createPlanItemInstanceQuery().caseInstanceId(caseInstance.getId()).count();
+        assertEquals(createCaseInstanceEndedErrorMessage(caseInstance, count), 0, count);
+        assertEquals("Runtime case instance found", 0, cmmnRuntimeService.createCaseInstanceQuery().caseInstanceId(caseInstance.getId()).count());
+        assertEquals(1, cmmnHistoryService.createHistoricCaseInstanceQuery().caseInstanceId(caseInstance.getId()).finished().count());
+    }
+
+    protected String createCaseInstanceEndedErrorMessage(CaseInstance caseInstance, long count) {
+        String errorMessage = "Plan item instances found for case instance: ";
+        if (count != 0) {
+            List<PlanItemInstance> planItemInstances = cmmnRuntimeService.createPlanItemInstanceQuery().caseInstanceId(caseInstance.getId()).list();
+            String names = planItemInstances.stream()
+                .map(planItemInstance -> planItemInstance.getName() + "(" + planItemInstance.getPlanItemDefinitionType() + ")")
+                .collect(Collectors.joining(", "));
+            errorMessage += names;
+        }
+        return errorMessage;
+    }
+
+    protected void assertCaseInstanceEnded(CaseInstance caseInstance, int nrOfExpectedMilestones) {
+        assertCaseInstanceEnded(caseInstance);
+        assertEquals(0, cmmnRuntimeService.createMilestoneInstanceQuery().milestoneInstanceCaseInstanceId(caseInstance.getId()).count());
+        assertEquals(nrOfExpectedMilestones, cmmnHistoryService.createHistoricMilestoneInstanceQuery().milestoneInstanceCaseInstanceId(caseInstance.getId()).count());
+    }
+    
+    protected void assertCaseInstanceNotEnded(CaseInstance caseInstance) {
+        assertTrue("Found no plan items for case instance", cmmnRuntimeService.createPlanItemInstanceQuery().caseInstanceId(caseInstance.getId()).count() > 0);
+        assertTrue("No runtime case instance found", cmmnRuntimeService.createCaseInstanceQuery().caseInstanceId(caseInstance.getId()).count() > 0);
+        assertNull("Historical case instance is already marked as ended", cmmnHistoryService.createHistoricCaseInstanceQuery().caseInstanceId(caseInstance.getId()).singleResult().getEndTime());
+    }
+    
+    protected void waitForJobExecutorToProcessAllJobs() {
+        CmmnJobTestHelper.waitForJobExecutorToProcessAllJobs(cmmnEngineConfiguration, 10000L, 200L, true);
+    }
+    
+    protected void waitForAsyncHistoryExecutorToProcessAllJobs() {
+        CmmnJobTestHelper.waitForAsyncHistoryExecutorToProcessAllJobs(cmmnEngineConfiguration, 10000L, 200L, true);
+    }
+
+}
