+++ modules/flowable-cmmn-engine/src/test/resources/flowable.cmmn.cfg.xml
@@ -5,11 +5,23 @@
 	xsi:schemaLocation="http://www.springframework.org/schema/beans   http://www.springframework.org/schema/beans/spring-beans.xsd">
 
 	<bean id="cmmnEngineConfiguration" class="org.flowable.cmmn.engine.impl.cfg.StandaloneInMemCmmnEngineConfiguration">
-	    <property name="jdbcUrl" value="jdbc:h2:mem:flowable;DB_CLOSE_DELAY=1000;MVCC=TRUE" />
-        <property name="jdbcDriver" value="org.h2.Driver" />
-        <property name="jdbcUsername" value="sa" />
+
+        <!-- cockroachdb specific -->
+        <property name="bulkInsertEnabled" value="true" />
+        <property name="databaseType" value="cockroachdb" />
+        <property name="databaseSchema" value="public" /> <!-- CRDB supports 'public' only currently -->
+
+        <property name="jdbcUrl" value="jdbc:postgresql://127.0.0.1:26257/flowable?sslmode=disable" />
+        <property name="jdbcDriver" value="org.postgresql.Driver" />
+        <property name="jdbcUsername" value="flowable" />
         <property name="jdbcPassword" value="" />
-        <property name="databaseSchemaUpdate" value="true" />
+        <property name="databaseSchemaUpdate" value="drop-create" />
+
+	    <!--<property name="jdbcUrl" value="jdbc:h2:mem:flowable;DB_CLOSE_DELAY=1000;MVCC=TRUE" />-->
+        <!--<property name="jdbcDriver" value="org.h2.Driver" />-->
+        <!--<property name="jdbcUsername" value="sa" />-->
+        <!--<property name="jdbcPassword" value="" />-->
+        <!--<property name="databaseSchemaUpdate" value="true" />-->
         
         <property name="asyncExecutor" ref="asyncExecutor" />
         <property name="asyncExecutorActivate" value="false" />
