+++ modules/flowable-cmmn-engine/src/test/resources/log4j.properties
@@ -1,4 +1,4 @@
-log4j.rootLogger=DEBUG, CA
+log4j.rootLogger=INFO, CA
 
 # ConsoleAppender
 log4j.appender.CA=org.apache.log4j.ConsoleAppender
@@ -12,4 +12,4 @@ log4j.logger.org.springframework=INFO
 log4j.logger.liquibase=INFO
 
 log4j.logger.org.flowable.task.service.impl.persistence=INFO
-log4j.logger.org.flowable.cmmn.engine.impl.agenda=DEBUG
+log4j.logger.org.flowable.cmmn.engine.impl.agenda=INFO
