+++ modules/flowable-eventsubscription-service/src/main/resources/org/flowable/eventsubscription/service/db/create/flowable.cockroachdb.create.eventsubscription.sql
@@ -1,0 +1,23 @@
+create table ACT_RU_EVENT_SUBSCR (
+    ID_ varchar(64) not null,
+    REV_ integer,
+    EVENT_TYPE_ varchar(255) not null,
+    EVENT_NAME_ varchar(255),
+    EXECUTION_ID_ varchar(64),
+    PROC_INST_ID_ varchar(64),
+    ACTIVITY_ID_ varchar(64),
+    CONFIGURATION_ varchar(255),
+    CREATED_ timestamp not null,
+    PROC_DEF_ID_ varchar(64),
+    SUB_SCOPE_ID_ varchar(64),
+    SCOPE_ID_ varchar(64),
+    SCOPE_DEFINITION_ID_ varchar(64),
+    SCOPE_TYPE_ varchar(64),
+    TENANT_ID_ varchar(255) default '',
+    primary key (ID_)
+);
+
+create index ACT_IDX_EVENT_SUBSCR_CONFIG_ on ACT_RU_EVENT_SUBSCR(CONFIGURATION_);
+create index ACT_IDX_EVENT_SUBSCR on ACT_RU_EVENT_SUBSCR(EXECUTION_ID_);
+
+insert into ACT_GE_PROPERTY values ('eventsubscription.schema.version', '6.5.0.0', 1);
