+++ modules/flowable-eventsubscription-service/src/main/resources/org/flowable/eventsubscription/service/db/drop/flowable.cockroachdb.drop.eventsubscription.sql
@@ -1,0 +1,1 @@
+drop table if exists ACT_RU_EVENT_SUBSCR cascade;
